Extract perl-dap configuration microcrate
