+++ Cargo.toml
@@ -102,6 +102,7 @@ members = [
     "crates/perl-dead-code",
     "crates/perl-dap-breakpoint",
     "crates/perl-dap-eval",
+    "crates/perl-dap-config",
     "crates/perl-dap-platform",
     "crates/perl-dap-command-args",
     "crates/perl-dap-shell",
@@ -216,6 +217,7 @@ allow = [
     # Tier 5 — DAP components
     "perl-dap-breakpoint",
     "perl-dap-eval",
+    "perl-dap-config",
     "perl-dap-command-args",
     "perl-dap-platform",
     "perl-dap-shell",
@@ -338,6 +340,7 @@ perl-dap = { path = "crates/perl-dap", version = "0.11.0" }
 perl-dead-code = { path = "crates/perl-dead-code", version = "0.11.0" }
 perl-dap-breakpoint = { path = "crates/perl-dap-breakpoint", version = "0.11.0" }
 perl-dap-eval = { path = "crates/perl-dap-eval", version = "0.11.0" }
+perl-dap-config = { path = "crates/perl-dap-config", version = "0.11.0" }
 perl-dap-platform = { path = "crates/perl-dap-platform", version = "0.11.0" }
 perl-dap-value = { path = "crates/perl-dap-value", version = "0.11.0" }
 perl-dap-command-args = { path = "crates/perl-dap-command-args", version = "0.11.0" }
