+++ crates/perl-dap-config/Cargo.toml
@@ -1,0 +1,27 @@
+[package]
+name = "perl-dap-config"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+authors.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+description = "Launch/attach configuration parsing and validation for perl-dap"
+readme = "README.md"
+documentation = "https://docs.rs/perl-dap-config"
+keywords = ["perl", "dap", "debug", "configuration", "launch"]
+categories = ["development-tools::debugging", "command-line-interface"]
+include = ["src/**", "Cargo.toml", "README.md", "LICENSE*"]
+
+[lib]
+name = "perl_dap_config"
+path = "src/lib.rs"
+
+[dependencies]
+anyhow = "1.0.102"
+serde = { version = "1.0.228", features = ["derive"] }
+serde_json = "1.0.149"
+
+[lints]
+workspace = true
