+++ crates/perl-dap-config/README.md
@@ -1,0 +1,11 @@
+# perl-dap-config
+
+Standalone SRP microcrate for Perl DAP launch/attach configuration models, validation,
+and editor snippet generation.
+
+## API
+
+- `LaunchConfiguration`
+- `AttachConfiguration`
+- `create_launch_json_snippet()`
+- `create_attach_json_snippet()`
